Use Uint32Array to Uint8Array to correctly encode length int32
Setting the value in the Uint8Array with the Uint32Array array
here is only setting the first byte of the array, even if
the length is greater than 256. This causes sync requests
to fail to deserialize the JSON buffer for long responses from
SQLite.

Using the Uint32Array buffer to populate a Uint8Array before
setting in the response buffer encodes longer integers correctly.